Reduced the steps needed to create a new practice. When you create a new practice session the steps are 1) choose a problem, click [next] 2) choose a language and test-framework, click [next] 3) click [confirm] 4) click [enter the exercise] This is too many clicks. So - Got rid of 3) and 4) - The [next] button for 2) can become [ok] - Clicking [ok] takes you straight to the kata/edit/ID URL - In a new tab/window - Without having to make an additional 'confirmation' click
Removed confirm.erb since it is not used anymore after "Reduced the steps needed to create a new practice."

The OK button must be an HREF to avoid popup block.
The kata must appear on a new window/tab, so the the user cannot click on the back button. Exercise progress could be lost.